+++ flutter_onestore_inapp/example/lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_onestore_inapp/flutter_onestore_inapp.dart';
 import 'package:provider/provider.dart';
 
 import 'purchase_view_model.dart';