+++ flutter_onestore_inapp/lib/src/onestore_base.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/services.dart';
+import 'onestore_channel.dart';
+
+
+const baseChannel = MethodChannel('${OneStoreChannel.rootChannel}/base');
+
+